# ElectricCarsServer/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "App for serving autoparts requests",
        "title": "Autoparts App",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/autoparts": {
            "put": {
                "description": "Update Autopart by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autoparts"
                ],
                "summary": "Update Autopart by admin",
                "parameters": [
                    {
                        "description": "updated autopart",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Autopart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Autopart"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Add Autopart by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autoparts"
                ],
                "summary": "Add Autopart by admin",
                "parameters": [
                    {
                        "description": "New autopart",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Autopart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/autoparts/add-to-assembly": {
            "post": {
                "description": "Add To Assembly by client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autoparts"
                ],
                "summary": "Add To Assembly by client",
                "parameters": [
                    {
                        "description": "Add To Assembly autopart",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.AddToAssemblyID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/autoparts/get-al": {
            "get": {
                "description": "Autoparts List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autoparts"
                ],
                "summary": "Autoparts List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query string to filter autoparts by name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.AutopartList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/autoparts/upload-image": {
            "post": {
                "description": "Add Image to autopart by admin",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autoparts"
                ],
                "summary": "Add Image to autopart by admin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "autopart ID",
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/autoparts_assembly": {
            "put": {
                "description": "Update Count Autopart Assembly by client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autopart_Assembly"
                ],
                "summary": "Update Count Autopart Assembly by v",
                "parameters": [
                    {
                        "description": "Delete Autopart From Assembly",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Autopart_Assembly"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated successfully",
                        "schema": {
                            "$ref": "#/definitions/ds.Autopart_Assembly"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete Autopart From Assembly by client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autopart_Assembly"
                ],
                "summary": "Delete Autopart From Assembly by client",
                "parameters": [
                    {
                        "description": "Delete Autopart From Assembly",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Autopart_Assembly"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/autoparts{id}": {
            "get": {
                "description": "Autopart By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autoparts"
                ],
                "summary": "Autopart By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Autoaprt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Autopart"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete Autopart by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autoparts"
                ],
                "summary": "Delete Autopart by admin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Autoaprt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.AddToAssemblyID": {
            "type": "object",
            "properties": {
                "autopart": {
                    "$ref": "#/definitions/ds.AutopartDetails"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "ds.Autopart": {
            "type": "object",
            "properties": {
                "autopart_id": {
                    "type": "integer"
                },
                "brand": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "ds.AutopartDetails": {
            "type": "object",
            "properties": {
                "autopart_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "ds.AutopartList": {
            "type": "object",
            "properties": {
                "autoparts_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Autopart"
                    }
                },
                "draft_id": {
                    "type": "integer"
                }
            }
        },
        "ds.Autopart_Assembly": {
            "type": "object",
            "properties": {
                "autopart_factory_id": {
                    "type": "integer"
                },
                "autopart_id": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "factory_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}